Update init_server.sh
add download check

+++ init_server.sh
@@ -19,20 +19,25 @@ then
  
  echo download bedrock-server-$VERSION.zip...
  wget -q --output-document=bedrock-server.zip https://www.minecraft.net/bedrockdedicatedserver/bin-linux/bedrock-server-$VERSION.zip
- echo saved in /data/bedrock-server.zip
-
- echo unzip bedrock-server.zip...
- unzip -oq bedrock-server.zip
- echo files unzipped in /data
-
- echo remove bedrock-server.zip...
- rm -f bedrock-server.zip
- echo bedrock-server.zip removed
+ if [ `wc -c <bedrock-server.zip` -ge 1000 ]
+ then
+  echo saved in /data/bedrock-server.zip
  
- echo get more rights to executable
- chmod 777 bedrock_server
+  echo unzip bedrock-server.zip...
+  unzip -oq bedrock-server.zip
+  echo files unzipped in /data
  
- echo $VERSION > $VERSION_FILE
+  echo remove bedrock-server.zip...
+  rm -f bedrock-server.zip
+  echo bedrock-server.zip removed
+  
+  echo get more rights to executable
+  chmod 777 bedrock_server
+  
+  echo $VERSION > $VERSION_FILE
+ else
+  echo download bedrock-server-$VERSION.zip failed
+ fi
 fi
 
 echo start bedrock_server
